Pass CacheType around everywhere, so that we can have meaningful error messages
